backend/requirements.txt
django
djangorestframework
mongoengine
python-dotenv
djangorestframework-simplejwt
requests
beautifulsoup4
scikit-learn
python-decouple
celery
django-celery-beat
django-cors-headers
fcm-django
django-filter
django-redis
django-extensions
django-crispy-forms

